Property bases test for conversions

+++ romans/cases.go
@@ -1,7 +1,7 @@
 package romans
 
 var cases = []struct {
-	Arabic int
+	Arabic uint16
 	Roman  string
 }{
 	{Arabic: 1, Roman: "I"},
